feat: widen HP, MP and WeaponDamage to long for dimension scaling
Abyss scaling formulas overflow int.MaxValue from dimension 3 (boss-tier)
and dimension 5 (everything). Widen AttributeSet.HP/MP, all WeaponDamage
paths, and the DB reads that feed them to long. SP stays int.

Audit list outcomes (all 18 rows):
  1  AttributeSet.HP/.MP               int -> long. DONE
  2  operator+ / operator- (HP,MP)     now long arithmetic; no cast. VERIFIED
  3  operator* HP/MP (:180,181)        (int)Math.Ceiling -> (long)Math.Ceiling. DONE
  4  Clone (:75,76)                    copies long -> long; not narrowing. VERIFIED
  5  NPCTemplate.WeaponDamage          int -> long. DONE
  6  NPC.WeaponDamage (:302)           int -> long. DONE
  7  ICharacter.WeaponDamage (:88)     int -> long. DONE
  8  Player.WeaponDamage (:1572) +     int -> long (return widens); item damage
     Inventory.GetWeaponDamage         stays int (items not dimension-scaled). DONE
  9  Pet.WeaponDamage (Pet.cs:76)      override int -> long. DONE
 10  Pet DB read (Pet.cs:265)          Convert.ToInt32 -> Convert.ToInt64. DONE
 11  Pet INSERT/UPDATE (:361,:432)     string concat; unchanged, long formats
                                       identically. VERIFIED
 12  NPCHandler DB reads (:54,81,82)   Convert.ToInt32 -> Convert.ToInt64. DONE
 13  Player DB reads player_hp/mp      Convert.ToInt32 -> Convert.ToInt64. DONE
 14  NPC.Attack damage (NPC.cs:1370)   double accumulator; long -> double implicit,
                                       unchanged. VERIFIED
 15  Player.Attack (1543-1549)         double accumulator; unchanged; WeaponDamage
                                       == 1 sentinel still reads correctly (long vs
                                       int literal 1). VERIFIED
 16  SpellEffect formula symbols       Dictionary<string, decimal>; long -> decimal
      (:1274,:1284)                    implicit, unchanged. VERIFIED
 17  Packets with HP/MP/damage         string concat; verified no int.Parse
                                       round-trip. Packets.cs:102/183/211/409/410
                                       cast only the 0-100 HP% ratio, not raw HP.
                                       Unchanged. VERIFIED
 18  Window.cs:285 pet weapon damage   string concat; unchanged. VERIFIED

Additional Convert.ToInt32 -> ToInt64 sites caught by the plan's grep
`Convert.ToInt32 | grep -iE hp|mp|weapon_damage` (same columns/fields feeding
the widened values; ToInt32 would throw OverflowException past int.MaxValue):
ClassHandler.cs:85,86 (class player_hp/mp), SpellHandler.cs:87,88 (spell
effect hp/mp), Pet.cs:230,231 (pet_hp/mp), ItemHandler.cs:77,78 (item
template player_hp/mp).

Left int on purpose (not dimension-scaled / not HP/MP/WeaponDamage values):
ItemTemplate.WeaponDamage, Item.WeaponDamage/TotalWeaponDamage,
IItem.WeaponDamage, regen fields, spell HP/MP costs, HP% packet ratio casts.

+++ Goose/AttributeSet.cs
@@ -11,8 +11,8 @@ namespace Goose
      */
     public class AttributeSet
     {
-        public int HP { get; set; }
-        public int MP { get; set; }
+        public long HP { get; set; }
+        public long MP { get; set; }
         public int SP { get; set; }
         public decimal HPPercentRegen { get; set; }
         public int HPStaticRegen { get; set; }
@@ -177,8 +177,8 @@ namespace Goose
 
             AttributeSet temp = new AttributeSet();
 
-            temp.HP = (int)Math.Ceiling(a1.HP * multiplier);
-            temp.MP = (int)Math.Ceiling(a1.MP * multiplier);
+            temp.HP = (long)Math.Ceiling(a1.HP * multiplier);
+            temp.MP = (long)Math.Ceiling(a1.MP * multiplier);
             temp.SP = (int)Math.Ceiling(a1.SP * multiplier);
             temp.HPPercentRegen = a1.HPPercentRegen * multiplier;
             temp.HPStaticRegen = (int)Math.Ceiling(a1.HPStaticRegen * multiplier);
